Merged the mobi DRM checking to the master branch

+++ flint-toolwrappers/src/main/java/uk/bl/dpt/qa/flint/wrappers/Jhove1Wrapper.java
@@ -74,9 +74,12 @@ public class Jhove1Wrapper {
 	 */
 	public boolean isValid(File pFile) {
 		boolean ret = false;
+		
+		File temp = null;
 		try {
 		    LOGGER.warn("jhove staring");
-			File temp = File.createTempFile("jhove-output-", ".xml");
+			temp = File.createTempFile("jhove-output-", ".xml");
+
 			String[] inputs = new String[] { pFile.getAbsolutePath() };
 			jhove.dispatch(app, null, null, handler, temp.getAbsolutePath(), inputs);
 			
@@ -93,10 +96,10 @@ public class Jhove1Wrapper {
 				//System.out.println(temp.getAbsolutePath()+": "+status+", "+status.length());
 			}
             LOGGER.warn("jhove thinks it is: {}", status.toLowerCase());
-			temp.delete();
-
 		} catch (Exception e) {
 			LOGGER.warn("Caught exception: {}", e);
+		} finally {
+		    if (temp != null) temp.delete();
 		}
 
 		return ret;
